requirements.txt
Django==2.2.24
djangorestframework==3.12.4
drf-writable-nested==0.1.0
pytest==3.10.1
pytest-django==3.4.8
pytest-cov==2.6.1
codecov==2.1.13